+++ src/main/java/com/example/originmodstudy/client/OriginModStudyClient.java
@@ -9,5 +9,6 @@ public class OriginModStudyClient implements ClientModInitializer {
 	@Override
 	public void onInitializeClient() {
 		EntityRendererRegistry.register(ModEntities.THROWN_HARPY_JAVELIN, ThrownHarpyJavelinRenderer::new);
+		EntityRendererRegistry.register(ModEntities.THROWN_SILK_NET, ThrownSilkNetRenderer::new);
 	}
 }